File download and upload. To doo:  ProgressBar in info window;  Replace window with equals files  Delete button

+++ cloud-client/src/main/java/gb/project/cloud/client/ClientController.java
@@ -2,12 +2,10 @@ package gb.project.cloud.client;
 
 import gb.project.cloud.client.dialog.*;
 import javafx.application.Platform;
+import javafx.concurrent.Task;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
-import javafx.scene.control.Label;
-import javafx.scene.control.ListView;
-import javafx.scene.control.MenuItem;
-import javafx.scene.control.TextField;
+import javafx.scene.control.*;
 import javafx.scene.paint.Color;
 
 import java.io.IOException;
@@ -38,22 +36,34 @@ public class ClientController implements Initializable {
     private MenuItem close;
     @FXML
     private MenuItem enter;
+    @FXML
+    private ProgressBar progressBar;
+    @FXML
+    private Label progressLabel;
 
     @FXML
     private void upload() {
         String filename = clientView.getSelectionModel().getSelectedItem();
         if (connect) {
-            clientNet.upload(clientDir.resolve(filename));
+            if (clientDir.resolve(filename).toFile().isFile()) {
+                clientNet.upload(clientDir.resolve(filename));
+            } else {
+                messageDialog("Ошибка", "Выберете файл");
+            }
         } else {
-            Path selected = clientDir.resolve(filename);
-            Path destFile = Paths.get(clientDir2.normalize().toAbsolutePath().toString(),
-                    selected.getFileName().toString());
-            try {
-                Files.copy(selected, destFile,
-                        StandardCopyOption.REPLACE_EXISTING);
-                updateClientView();
-            } catch (IOException ex) {
-                ex.printStackTrace();
+            if (clientDir.resolve(filename).toFile().isFile()) {
+                Path selected = clientDir.resolve(filename);
+                Path destFile = Paths.get(clientDir2.normalize().toAbsolutePath().toString(),
+                        selected.getFileName().toString());
+                try {
+                    Files.copy(selected, destFile,
+                            StandardCopyOption.REPLACE_EXISTING);
+                    updateClientView();
+                } catch (IOException ex) {
+                    ex.printStackTrace();
+                }
+            } else {
+                messageDialog("Ошибка", "Выберете файл");
             }
         }
     }
@@ -110,73 +120,13 @@ public class ClientController implements Initializable {
         //обработчик окна файлов компьютера
         clientView.setOnMouseClicked(e -> {
             if (e.getClickCount() == 2) {
-                String item = clientView.getSelectionModel().getSelectedItem();
-                //если в окне нажали на пустую строку
-                if (item == null) {
-                    return;
-                }
-                if (item.equals("...")) {
-                    clientDir = clientDir.getParent();
-                    updateClientView();
-                } else {
-                    Path selected = clientDir.resolve(item);
-                    if (selected.toFile().isDirectory()) {
-                        clientDir = selected;
-                        updateClientView();
-                    }
-                }
+                clientView2Click();
             }
         });
         //обработчик окна файлов сервера (если нет соединения то сервер = компьютеру)
         serverView.setOnMouseClicked(e -> {
             if (e.getClickCount() == 2) {
-                String item = serverView.getSelectionModel().getSelectedItem();
-                //если в окне нажали на пустую строку
-                if (item == null) {
-                    return;
-                }
-                if (item.equals("...")) {
-                    //если не подключен к серверу
-                    if (!connect) {
-                        clientDir2 = clientDir2.getParent();
-                        List<String> names2;
-                        if (clientDir2.getParent() != null) {
-                            names2 = new LinkedList<>();
-                            names2.add("...");
-                            names2.addAll(List.of(Objects.requireNonNull(clientDir2.toFile().list())));
-                        } else {
-                            names2 = List.of(Objects.requireNonNull(clientDir2.toFile().list()));
-                        }
-                        updateServerView("this computer", names2, clientDir2.toAbsolutePath().normalize().toString());
-                    }
-                    //при подключенном
-                    else {
-                        clientNet.directory(item);
-                    }
-                } else {
-                    //если не подключен к серверу
-                    if (!connect) {
-                        Path selected = clientDir2.resolve(item);
-                        if (selected.toFile().isDirectory()) {
-                            clientDir2 = selected;
-                            List<String> names2;
-                            if (clientDir2.getParent() != null) {
-                                names2 = new LinkedList<>();
-                                names2.add("...");
-                                names2.addAll(List.of(Objects.requireNonNull(clientDir2.toFile().list())));
-                            } else {
-                                names2 = List.of(Objects.requireNonNull(clientDir2.toFile().list()));
-                            }
-                            updateServerView("this computer",
-                                    names2,
-                                    clientDir2.toAbsolutePath().normalize().toString());
-                        }
-                    }
-                    //при подключенном
-                    else {
-                        clientNet.directory(item);
-                    }
-                }
+                serverView2Click();
             }
         });
     }
@@ -279,10 +229,90 @@ public class ClientController implements Initializable {
 
     public void enterToServer() {
         String[] a = userDataWindow("!", "Enter userdata");
-        clientNet.login(a[0],a[1]);
+        clientNet.login(a[0], a[1]);
     }
 
     public void closeConnection() {
         setDisconnect();
     }
+
+    public void progressCopy(String type, long current, long size) {
+        if ("upload".equals(type)) {
+            System.out.println("Msg: " + type + " get " + current + " size " + size);
+            progressBar.setProgress(current / size);
+            clientNet.pathUpload();
+        }
+        if ("get".equals(type)) {
+            progressBar.setProgress(current / size);
+        }
+    }
+
+    private void clientView2Click() {
+        String item = clientView.getSelectionModel().getSelectedItem();
+        //если в окне нажали на пустую строку
+        if (item == null) {
+            return;
+        }
+        if (item.equals("...")) {
+            clientDir = clientDir.getParent();
+            updateClientView();
+        } else {
+            Path selected = clientDir.resolve(item);
+            if (selected.toFile().isDirectory()) {
+                clientDir = selected;
+                updateClientView();
+            }
+        }
+    }
+
+    private void serverView2Click() {
+        String item = serverView.getSelectionModel().getSelectedItem();
+        //если в окне нажали на пустую строку
+        if (item == null) {
+            return;
+        }
+        if (item.equals("...")) {
+            //если не подключен к серверу
+            if (!connect) {
+                clientDir2 = clientDir2.getParent();
+                List<String> names2;
+                if (clientDir2.getParent() != null) {
+                    names2 = new LinkedList<>();
+                    names2.add("...");
+                    names2.addAll(List.of(Objects.requireNonNull(clientDir2.toFile().list())));
+                } else {
+                    names2 = List.of(Objects.requireNonNull(clientDir2.toFile().list()));
+                }
+                updateServerView("this computer", names2, clientDir2.toAbsolutePath().normalize().toString());
+            }
+            //при подключенном
+            else {
+                clientNet.directory(item);
+            }
+        } else {
+            //если не подключен к серверу
+            if (!connect) {
+                Path selected = clientDir2.resolve(item);
+                if (selected.toFile().isDirectory()) {
+                    clientDir2 = selected;
+                    List<String> names2;
+                    if (clientDir2.getParent() != null) {
+                        names2 = new LinkedList<>();
+                        names2.add("...");
+                        names2.addAll(List.of(Objects.requireNonNull(clientDir2.toFile().list())));
+                    } else {
+                        names2 = List.of(Objects.requireNonNull(clientDir2.toFile().list()));
+                    }
+                    updateServerView("this computer",
+                            names2,
+                            clientDir2.toAbsolutePath().normalize().toString());
+                }
+            }
+            //при подключенном
+            else {
+                clientNet.directory(item);
+            }
+        }
+    }
+
 }
