+++ cloud-client/src/main/java/gb/project/cloud/client/service/MessageResponse.java
@@ -5,6 +5,7 @@ import gb.project.cloud.objects.*;
 import javafx.application.Platform;
 
 import java.nio.file.Files;
+import java.nio.file.StandardOpenOption;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -12,7 +13,9 @@ public class MessageResponse {
     private final static Map<MessageType, ServiceMessage> responseMap = new HashMap<>();
 
     public MessageResponse(ClientController client, String host, int port) {
+
         //AUTH
+
         responseMap.put(MessageType.AUTH, (ctx, cm) -> {
             AuthMessage am = (AuthMessage) cm;
             if (am.getTypeAuth() == 0) {
@@ -22,16 +25,46 @@ public class MessageResponse {
                 Platform.runLater(() -> client.messageDialog("Error", "Wrong username or password"));
             }
         });
+
         //FILE
+
         responseMap.put(MessageType.FILE, (ctx, cm) -> {
             FileMessage fm = (FileMessage) cm;
-            Files.write(client.getClientDir().resolve(fm.getName()), fm.getBytes());
-            client.updateClientView();
+            if (client.getClientDir().resolve(fm.getName()).toFile().exists()) {
+                Files.write(
+                        client.getClientDir().resolve(fm.getName()),
+                        fm.getBytes(),
+                        StandardOpenOption.APPEND
+                );
+                client.progressCopy(
+                        "get",
+                        client.getClientDir().resolve(fm.getName()).toFile().length(),
+                        fm.getSize()
+                );
+            } else {
+                Files.write(client.getClientDir().resolve(fm.getName()), fm.getBytes());
+                client.progressCopy(
+                        "get",
+                        client.getClientDir().resolve(fm.getName()).toFile().length(),
+                        fm.getSize()
+                );
+            }
+
+            if (client.getClientDir().resolve(fm.getName()).toFile().length() == fm.getSize()) {
+                client.updateClientView();
+            }
         });
+
         //FILE_REQUEST
+
         responseMap.put(MessageType.LIST, (ctx, cm) -> {
             ListMessage lm = (ListMessage) cm;
             Platform.runLater(() -> client.updateServerView(host + ":" + port, lm.getFiles(), lm.getPath()));
+        });
+
+        responseMap.put(MessageType.PATH_GET, (ctx, cm) -> {
+            PathFileGet gp = (PathFileGet) cm;
+            client.progressCopy(gp.getTypeMes(), gp.getGatedBytes(), gp.getSizeFile());
         });
     }
